requirements.txt
joblib==1.4.2
keras==3.5.0
matplotlib==3.9.2
numpy==2.1.1
Pillow==10.4.0
scikit_learn==1.5.2
seaborn==0.13.2
tensorflow==2.17.0
